+++ src/utils/refToId.ts
@@ -1,4 +1,15 @@
-import { parseRef } from '.';
+/**
+ * Parses OpenApi ref:
+ * "#/components/schema/Foo" --> "components/schema/Foo"
+ */
+function parseRef(ref: string): string {
+  if (!ref.startsWith('#/')) {
+    throw new Error(`[openapi-ts-json-schema] Unsupported ref value: "${ref}"`);
+  }
+
+  const refPath = ref.replace('#/', '');
+  return refPath;
+}
 
 /**
  * Generate an internal schema ID from a given schema ref:
