+++ test/unit/refToId.test.ts
@@ -1,18 +1,18 @@
 import { describe, it, expect } from 'vitest';
-import { parseRef } from '../../src/utils';
+import { refToId } from '../../src/utils';
 
-describe('parseRef', () => {
+describe('refToId', () => {
   describe('Valid ref', () => {
     it('returns, ref path', () => {
-      const actual = parseRef('#/components/schemas/Foo');
-      const expected = 'components/schemas/Foo';
+      const actual = refToId('#/components/schemas/Foo');
+      const expected = '/components/schemas/Foo';
       expect(actual).toBe(expected);
     });
   });
 
   describe('Invalid ref', () => {
     it('throws error', () => {
-      expect(() => parseRef('/components/schemas/Foo')).toThrow(
+      expect(() => refToId('/components/schemas/Foo')).toThrow(
         new Error(
           `[openapi-ts-json-schema] Unsupported ref value: "/components/schemas/Foo"`,
         ),
